@@ -1,19 +1,20 @@
 plugins {
 	id 'java'
-	id 'org.springframework.boot' version '2.7.9'
-	id 'io.spring.dependency-management' version '1.1.3'
+	id 'org.springframework.boot' version '3.5.3'
+	id 'io.spring.dependency-management' version '1.1.7'
 }
+
+group = 'com.example'
+version = '0.0.1-SNAPSHOT'
+
 java {
-	sourceCompatibility = '11'
+	toolchain {
+		languageVersion = JavaLanguageVersion.of(17)
+	}
 }
-group 'org.example'
-version '1.0-SNAPSHOT'
 
 repositories {
 	mavenCentral()
-}
-jar {
-	enabled = false
 }
 
 dependencies {
