@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
@@ -38,4 +38,15 @@ public class ProblemStats {
     @UpdateTimestamp
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
+
+
+    // 정답률을 Double로 반환하는 헬퍼 메서드
+    public Double getCorrectRateAsDouble() {
+        return correctRate != null ? correctRate.doubleValue() : 0.0;
+    }
+
+    // 정답률을 설정하는 헬퍼 메서드
+    public void setCorrectRateFromDouble(Double rate) {
+        this.correctRate = rate != null ? BigDecimal.valueOf(rate) : BigDecimal.ZERO;
+    }
 }
