@@ -1,5 +1,6 @@
 package com.company.edu.entity;
 
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,29 +8,36 @@ import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import jakarta.persistence.*;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "multiple_choice_answers")
+@Table(name = "semesters")
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class MultipleChoiceAnswer {
+public class Semesters {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    @OneToOne
-    @JoinColumn(name = "problem_id", nullable = false)
-    private Problem problem;
+    @Enumerated(EnumType.STRING)
+    @Column(name = "grade_level", nullable = false)
+    private GradeLevel gradeLevel;
 
-    @Column(name = "correct_option_number", nullable = false)
-    private Integer correctOptionNumber;
 
-    @Column(name = "total_options", nullable = false)
-    private Integer totalOptions;
+    @Column(name = "semester_name", nullable = false)
+    private String semesterName;
+
+    @Column(name = "semester_code", nullable = false)
+    private String semesterCode;
+
+    @Column(name = "curriculum_year", nullable = false)
+    private Integer curriculumYear;
+
+    @Column(name = "is_active")
+    private byte isActive;
 
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
@@ -38,4 +46,9 @@ public class MultipleChoiceAnswer {
     @UpdateTimestamp
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
+
+
+    public enum GradeLevel {
+        elementary, middle, high;
+    }
 }
